# .context/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: JobWise API
  description: |
    JobWise AI-powered job application assistant API.
    
    ## Features
    - Master resume profile management
    - Job discovery and management
    - AI-powered resume and cover letter generation (5-stage pipeline)
    - Document management and PDF generation
    - Real-time generation progress tracking
    
    ## Authentication
    - Development: API Key authentication
    - Production: JWT Bearer tokens with OAuth 2.0
    
    ## Rate Limiting
    - AI Generation: 10 requests/hour per user
    - Other endpoints: 100 requests/minute per user
    
  version: 1.0.0
  contact:
    name: JobWise Development Team
    email: dev@jobwise.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.jobwise.ai/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /generations/resume:
    post:
      tags:
        - AI Generation
      summary: Generate tailored resume
      description: |
        Generate AI-tailored resume using 5-stage pipeline:
        1. Job Analysis - Extract keywords and requirements (1500 tokens)
        2. Profile Compilation - Score and rank profile content (2000 tokens)  
        3. Document Generation - Create tailored resume content (3000 tokens)
        4. Quality Validation - Verify ATS compliance and factuality (1500 tokens)
        5. PDF Export - Generate professional PDF (0 tokens)
        
        Total token budget: 8000 tokens per generation
        Performance target: <30s (p50), <60s (p95)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeGenerationRequest'
            example:
              profile_id: "profile_456"
              job_id: "job_123"
              options:
                template: "modern"
                length: "one_page"
                focus_areas:
                  - "backend_development"
                  - "leadership"
                include_cover_letter: true
                custom_instructions: "Emphasize cloud architecture experience"
      responses:
        '201':
          description: Resume generation started
          headers:
            Location:
              description: URL to check generation status
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationResponse'
              example:
                generation_id: "gen_789"
                status: "pending"
                estimated_completion: "2024-10-18T10:32:00Z"
                progress:
                  current_stage: 0
                  total_stages: 5
                  percentage: 0
                profile_id: "profile_456"
                job_id: "job_123"
                created_at: "2024-10-18T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "rate_limit_exceeded"
                message: "Generation rate limit exceeded. Try again in 1 hour."
                retry_after: 3600
  /generations/cover-letter:
    post:
      tags:
        - AI Generation
      summary: Generate tailored cover letter
      description: Generate AI-tailored cover letter for specific job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoverLetterGenerationRequest'
      responses:
        '201':
          description: Cover letter generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationResponse'
  /generations/{generation_id}:
    get:
      tags:
        - AI Generation
      summary: Get generation status
      description: Check status and progress of AI generation process
      parameters:
        - $ref: '#/components/parameters/GenerationId'
      responses:
        '200':
          description: Generation status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationStatusResponse'
              examples:
                in_progress:
                  summary: Generation in progress
                  value:
                    generation_id: "gen_789"
                    status: "generating"
                    progress:
                      current_stage: 2
                      total_stages: 5
                      percentage: 40
                      stage_name: "Profile Compilation"
                      stage_description: "Scoring profile content by relevance"
                    tokens_used: 3500
                    estimated_completion: "2024-10-18T10:32:00Z"
                    created_at: "2024-10-18T10:30:00Z"
                    updated_at: "2024-10-18T10:31:00Z"
                completed:
                  summary: Generation completed
                  value:
                    generation_id: "gen_789"
                    status: "completed"
                    progress:
                      current_stage: 5
                      total_stages: 5
                      percentage: 100
                      stage_name: "PDF Export"
                      stage_description: "Completed"
                    result:
                      document_id: "doc_101"
                      ats_score: 0.87
                      match_percentage: 82
                      keyword_coverage: 0.91
                      pdf_url: "https://api.jobwise.ai/documents/doc_101/download"
                    tokens_used: 7850
                    generation_time: 28.5
                    created_at: "2024-10-18T10:30:00Z"
                    completed_at: "2024-10-18T10:31:28Z"
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - AI Generation
      summary: Cancel generation
      description: Cancel ongoing AI generation process
      parameters:
        - $ref: '#/components/parameters/GenerationId'
      responses:
        '204':
          description: Generation cancelled successfully
        '400':
          description: Cannot cancel completed generation
        '404':
          $ref: '#/components/responses/NotFound'

  # Document Management
...
  /documents/{document_id}/download:
    get:
      tags:
        - Documents
      summary: Download document PDF
      description: Download generated document as PDF file
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=\"John_Doe_Resume_TechCorp.pdf\""
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
        '404':
          $ref: '#/components/responses/NotFound'

  # POST /documents is handled above at /documents path
    
    DocumentId:
      name: document_id
      in: path
      required: true
      description: Document unique identifier
      schema:
        type: string
        format: uuid
...